Show which patches failed attestation
If one patch in the series fails attestation, show which one it is and
point out which part failed (commit metadata, commit message, or patch
content). At this time, this only works if it's not the *first* patch
that fails attestation, because otherwise we aren't able to locate the
attestation document in the first place. There's no simple solution
here, because otherwise we'd have to hit lore too many times, which
would be murder on large series.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>