Handle vN that only exists in the cover letter
There are special cases where we have:

cover: [PATCH v3 00/NN] Foo
 \- [PATCH 01/NN] Patch 1
  - [PATCH 02/NN] Patch 2

While the submitter *should* be setting "v3" in all patches, it's not
uncommon to have that inferred from the cover letter. When cover letter
is properly upthread from the patch that has an inferred v1 revision, we
can check the revision on the cover letter and fix the patch revision to
match.

Reported-by: Mark Brown <broonie@kernel.org>
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>