Properly handle blank lines in patches
We don't need to strip() patches, because a) it introduces a problem if
the patch has blank lines in it, and b) because it doesn't matter if
there's anything trailing at the end.

We still need better handling for obviously broken patches, but at least
valid ones shouldn't cause problems.

Reported-by: Geoff Levand <geoff@infradead.org>
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>