Use trailers from old revisions if no changes
If we have multiple versions of patch series in the same thread and
someone adds a follow-up trailer to a patch in v1, but doesn't do it for
the same *exact* patch in v2, then we carry over that trailer from v1 to
v2.

We use attestation hashes for this purpose, so all patch/metadata must
be exact in order for the trailer to be reused.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>