Recognize merged PRs when attempting auto-base
When we come across a pull request that doesn't list its base (i.e. not
generated using git-request-pull), we try to figure out the base
ourselves by doing merge-base. However, if the pull request is already
merged, then that is going to be a useless operation, so recognize this
situation and bail out early.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>