Loosen compatible release identifiers for install_requires
The install_requires entries use a compatible release operator.  As an
example, "requests~=2.24.0" maps to a requirement of ">= 2.24.0 and ==
2.24.*".  With the current version of requests (2.25.1), this leads to
a ContextualVersionConflict failure at runtime.

Allowing only Z to tick in version X.Y.Z seems unnecessarily strict
unless there are known problems with a particular release, and it
makes it more difficult for distributions to package b4.  Drop the
trailing digit from all of the version identifiers, allowing both Y
and Z to increase.

Signed-off-by: Kyle Meyer <kyle@kyleam.com>
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>