Reimplement attestation code one more time
Move end-to-end attestation code into its own library: patatt. See
https://git.kernel.org/pub/scm/utils/patatt/patatt.git/about/

It is included into b4 as a submodule, but you will need to init it
first:

git submodule update --init

This change significantly simplifies our attestation code, dropping
thousands of lines of rather hairy code. Notably, patatt-style
attestation is incompatible with previous attestation implementations
done directly in b4, but that's just as well -- we've always marked it
as "experimental" and the lack of adoption was proving that we weren't
on the right path.

Next to come is keyring management and documentation.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>