Implement trim_body support
When a message has a developer signature but is failing the signature
check, rerun it again with trim_body. If that passes, we know that the
signature is failing due to mailing list junk appended to the bottom of
the message. In that case, automatically trim the message body so we
have exactly what the developer attested and signed.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>