Account for in-body headers when trimming body
When we discover that a message can only be attested after we trim the
body, we *must* set the body to that version, otherwise an attacker
could append arbitrary content past the l= value boundary. We already do
this in the current form, but we weren't properly handing in-body
headers like From: and Subject: that are used to indicate to git the
patch author vs. committer.

This patch set fixes that and also streamlines a few other places where
we were already relying on git mailinfo calls.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>