Tentative suport for sending exploded series
I've been working on a way to automatically convert pull requests into
series, complete with mailing them out to arbitrary destinations. This
would allow folks to send a pull request to a dedicated list and have it
automatically converted into a well-formed series.

This is a tentative implementation that relies on git-send-email to do
most of the heavy lifting. I have misgivings about using git-send-email
for this purpose, but it does reduce the amount of duplicated code we
would have otherwise had to write, and allows us to hook into things
like tocmd/cccmd, etc.

For example, adding the following to your .git/config:

    [sendemail "autopr"]
        smtpserver = [your.server.here]
        smtpserverport = 587
        smtpencryption = tls
        smtpuser = [your-user]
        smtppass = [your-pass]
        transferEncoding = 8bit
        suppressFrom = yes
        confirm = never
        validate = no
        tocmd = "$(git rev-parse --show-toplevel)/scripts/get_maintainer.pl --norolestats --nol"
        cccmd = "$(git rev-parse --show-toplevel)/scripts/get_maintainer.pl --norolestats --nom"

This would allow doing the following:

    b4 pr -e -f "AutoPR Exploder <autopr@yourdomain.here>" -s autopr [--dry-run]

The pull request will be exploded into a patch series and sent to all
the proper destinations as returned by get_maintainer.pl. We construct
the message headers in a way that allow regular code review and "b4 am"
usage after the auto-exploded series is sent out.

If testing goes well, we'll implement this as a kernel.org service and
then hook a similar implementation via Gitlab/Github.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>