When deduping, prefer DKIM-validating messages
With newer lore.kernel.org and /all/, we get duplicate messages when
message bodies are different due to one of the messages passing through
a DKIM-compliant list, and another one through something that injects
in-body or in-subject junk. When dealing with duplicates, check both for
DKIM status and prefer the message that actually passes DKIM validation.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>