Document listid-preference config parameter
When we query /all/, we often get duplicates if a message was
crossposted to multiple lists and some of those lists altered the
message subject/body (e.g. legacy mailman, groups.io, etc). The
listid-preference parameter allows us to dedupe based on lists that are
least likely to mangle the messages.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>