Reimplement unicode control character logic
Instead of hardcoding the logic to bail at all presence of control
characters, soften it a bit to only error out when we see the presence
of Cfs in the absence of any other character from a non-latin character
set in the same line.

Additionally, show exactly where the Cfs were found when printing out
the error message.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>