mbox: query latest revision of patch series by patch subject
With get_extra_series([], base_msg=base_msg, direction=0, ...),
get the latest revision of series from public-inbox with a match to
base_msg subject.

This will be used when base_msg refer to a patch in a local mbox where
neither the revision nor the date nor author of the local patch is
relevant for the search.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>