Initial patchwork integration support
A lot of maintainers use patchwork alongside b4, to make it easier to
track patches and rely on some CI integration. This commit adds some
basic patchwork integration:

- on "b4 am", "b4 shazam", "b4 pr" we will mark the relevant patchwork
  entries as "Under Review"
- on "b4 ty" we can set these patches as "Accepted"
- on "b4 ty -d" we can set them as "Deferred"

To make it work, the following entries must be present in the repository
used with b4:

    [b4]
      pw-key = (your API token)
      pw-url = https://patchwork.kernel.org
      pw-project = (your project, e.g. linux-usb)
      pw-review-state = under-review
      pw-accept-state = accepted
      pw-discard-state = deferred

To get your patchwork API token, go to your patchwork profile page.

The pw-accept-state and pw-discard-state can be overridden using the
--pw-set-state flag to "b4 ty". E.g. if you wanted to mark the patches
as "Not applicable":

    b4 ty -d 5 --pw-set-state not-applicable

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>