ez-send: initial support for tagging sent revisions
After we sent off the revision, create the structure in the tree that
would allow us to go back to that revision regardless of what changes we
make to the working branch. This is tricky for the default "commit"
strategy, because we want to remove the cover letter from the history
and store its final version in the tag applied to the tip commit.

For this, we create a detached head, cherry-pick the range without the
cover, and then tag the resulting detached head.

Future work will make it possible to prep and send these as pull
requests.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>