send-receive: add some anti-spam protections
The goal of this service is to accept and send patches, nothing else.
It's not a replacement for an SMTP server, just a replacement for really
terrible SMTP servers that mangle patches. So, add some anti-spam
protections:

- only accept email that looks like patches or cover letters
- don't accept anything other than text/plain mail
- require that one of the to/cc addresses matches a predefined list of
  recognized mailing lists

Not a guarantee that this service won't get abused, but it's a start to
make sure that it won't be quite as tasty of a target.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>