Allow defining in-tree .b4-config for send- parameters
We want to be able to allow projects using b4 for patch submission to
define some toplevel parameters, such as where to send the series, or
which endpoint to use. We only use this for b4.send- options, and for
now don't override anything set in any other config files.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>