ez: reimplement extra prefix functionality
Previously, to add extra prefixes to the subject one had to pass them
every time as "b4 send --prefixes". This is awkward, because many
subsystems require that an identifier is added to all patches, such as
"netdev" or "bpf", so we really should have a more permanent way to add
prefixes to the series.

This reimplements the functionality via "b4 prep --set-prefixes", e.g.:

    b4 prep --set-prefixes netdev

Similarly, an RFC set of patches would get something like:

    b4 prep --set-refixes RFC netdev

To delete any extra prefixes:

    b4 prep --set-prefixes ''

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>