ez: implement b4 send --reflect mode
This implements a special testing mode that can be invoked with
--reflect. It will still populate the To/Cc headers with the destination
addresses, but will only email the address in "From:". This should be
particularly useful when using the web endpoint, because this allows
someone to check what their series will look like with the
post-processing that is applied by the endpoint (for DMARC purposes, for
example).

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>