ez: fix bugs when creating a prep-branch from single-patch series
We weren't properly setting the from-thread tracking variable when the
series that we used as our base didn't have a cover letter (such as, a
single-patch series).

Additionally, there was a regression introduced by rewriting the message
retrieval function that was ignoring the --from-thread parameter due to
a collision on the "msgid" keyword in argparse.

Reported-by: Mark Brown <broonie@kernel.org>
Link: https://msgid.link/Y4%2FhIf%2Bo%2Fe3AtBCD%40sirena.org.uk
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>