ez: add a large REFLECT MODE notice
It can be confusing and worrying to see the To: and Cc: headers
populated in a message sent with --reflect. Not everyone necessarily
knows that these headers aren't actually used for actual message
delivery and all that matters is what you give the remote server during
the "RCPT TO" stage of the SMTP connection.

Add a large notice to alert contributors about this, so they don't worry
that they accidentally sent something they didn't intend to send.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>