ez: avoid header wrapping when writing to file
Python's email module and git-send-email don't agree on what should
happen with headers of certain lengths. Python likes to wrap things as
follows:

    Subject:
     [PATCH v5 4/5] arm64: dts: rockchip: sync px30 DTSI with Linux kernel v6.1

This is kosher with RFC, but git-send-email has trouble parsing this
format (which is really a problem with git-send-email, but I also agree
that Python could be just breaking the line at the word boundary instead
of moving the entire thing to the next line.

This was introduced when we switched to the SMTP policy for dryrun
output, which enforces header lengths.

Reported-by: Quentin Schulz <quentin.schulz@theobroma-systems.com>
Link: https://msgid.link/b79c49b3-5815-7100-e99c-440e1840b3f9@theobroma-systems.com
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>