mbox: add --refetch command
One of the changes in 0.11 was making "b4 am" generate messages in the
same format as produced by "git format-patch" -- with minimal headers
and performing the in-body-From and header-from replacements as
necessary.

This command allows refetching all messages in a mailbox file so they
have their original headers and bodies.

Suggested-by: Mark Brown <broonie@kernel.org>
Link: https://msgid.link/Y7hb17cciwrXm4E1@sirena.org.uk
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>