ez: treat cover-letter as git commit msg
It is common to write the cover-letter in a merge commit message. For
this reason, it is recommended to wrap it at 72 chars and have a short
title. It is like a git commit message at the end while lines starting
with '#' will be ignored.

For this reason and to "trick" the text editor, the temp file is now
called COMMIT_EDITMSG like the one generated by 'git commit'.

At the end, the file didn't contain any restructured-text format.

While at it, the default template is also wrapped to 72 chars.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Link: https://msgid.link/20230112-cover-letter-72-chars-v1-1-69b15b1e3fef@tessares.net
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>