ez: get_series_details: correctly compute 'base_commit' for all strategies
In f0fad6e (ez: refactor based on initial feedback, 2022-07-20), support
was added for the 'branch-description' strategy, but
'get_series_details' was not changed to account for the fact that if the
cover letter is stored in the branch description, base_commit and
start_commit should be the same commit.

This was also missed when 9e95d52 (ez: implement tip-commit strategy,
2022-07-27) added support for the 'tip-commit' strategy and changed
get_series_details to adjust the end commit of the series.

Fix that by setting 'base_commit' to 'start_commit's parent only for the
'commit' strategy, and to 'start_commit' for other strategies.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>
Link: https://msgid.link/20230221-fix-base-commit-for-non-commit-strategies-v1-1-0e31f1333d76@gmail.com
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>