ez: allow sending all versions of a patch series in the same thread
Some projects prefer further iterations of a patch series to be sent in
the same thread as previous ones. Usually this means that the cover
letter of v2 is sent as a reply to the cover letter of v1, etc.

Add a new config option for 'b4 send', 'send-same-thread', defaulting to
False, and read its value in get_prep_branch_as_patches. When True, add
an 'In-Reply-To' header to the first mail in the series, referencing the
Message-ID of the previous iterations's cover letter.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>
Link: https://msgid.link/20230219-send-iterations-in-same-thread-v2-1-3ae4ac504e07@gmail.com
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>