ez: use full change-id when tagging sent revisions
I ran into a situation where I used the same branch name to send two
independent patch series. When this happens, we have a collision for
sent/tagname-vX, so avoid this by using the full change-id of the series
to make the unique tag. This change is backwards-compatible, so it will
check for both situations when performing compare-to or other
operations.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>