Fix: wrong From attribution on trim_body required for attestation
In very certain conditions, we lost the in-body From: and used the outer
header from:

- it needed to be a patatt-signed message
- you had to have the key for proper verification
- the mailing list server had to have appended some trailer junk

In this case, we reconstituted the body from the trimmed verification
results, but didn't properly put back the From: and Subject: data that
was stripped from it. It used to still work properly, until we
implemented the proper scissors handling in 0.10.

Reported-by: Palmer Dabbelt <palmer@dabbelt.com>
Link: https://msgid.link/mhng-2765f5e1-e3d9-4f36-82e2-1212e13b0cdd@palmer-ri-x1c9a
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>