shazam: warn when overriding base-commit with --merge-base
If the --merge-base argument is provided, base-commit information that
was provided by the submitter is ignored. This behaviour is intentional,
but warning the user that this has happened (and what the intended base
was) is helpful.

To be able to do this, base-commit can't be set from mergebase until
after parsing the submission and attempting to guess it.

Suggested-by: Palmer Dabbelt <palmer@rivosinc.com>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Link: https://msgid.link/20230330110909.355701-1-conor.dooley@microchip.com
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>