b4.sh: keep existing PYTHONPATH if set
for setups depending on PYTHONPATH being set to provide dependencies
e.g. requests, b4.sh overriding of the variable is too strong.

Set current directory and patatt's first but keep PYTHONPATH if it was
not empty.

While we are here, remove useless use of env: the shell does the job
just fine if the variable is set before exec

Signed-off-by: Dominique Martinet <asmadeus@codewreck.org>
Link: https://msgid.link/20230402-wrapper-keep-pythonpath-v1-1-e2bc5e556a67@codewreck.org
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>