mbox.py::make_am: simplify check for early return
In mbox.py::make_am, we check if the 'lser' LoreSeries returned by
lmbx.get_series is empty, and return early with a different critical
message depending on wether wantver was given or not.

Simplify the logic a bit by using a nested if instead of two subsequent
ones. This will make the following commit's diff simpler.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>
Link: https://msgid.link/20230415-am-cherry-pick-suggestion-v1-1-8951a2274256@gmail.com
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>